fixup refactor: rename retries to update retries
